New functionality (crop and resync)
Add ability to crop the windows area of the subtitle and to times by a resync (multiplicative) factor

The crop functionality is not fully tested as some path were never encountered during my tests, mostly the multiple windows\Composition Object, not fully tested is also the case in which the subtitle goes outside the new screen area, a warning is returned when this happens. Totally not implemented is the handling of the "Object Cropped Flag".